# public/swagger.yaml
openapi: 3.0.0
info:
  title: Memo API
  version: 1.0.0
  description: 메모 애플리케이션을 위한 API 문서

servers:
  - url: http://localhost:3000
    description: 개발 서버

paths:
  /api/memos:
    get:
      summary: 모든 메모 조회
      description: 생성일 기준 내림차순으로 정렬된 모든 메모를 조회합니다.
      responses:
        '200':
          description: 메모 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  memos:
                    type: array
                    items:
                      $ref: '#/components/schemas/Memo'
        '500':
          description: 서버 에러
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: 새 메모 생성
      description: 새로운 메모를 생성합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MemoInput'
      responses:
        '201':
          description: 메모 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: 생성된 메모의 ID
        '500':
          description: 서버 에러
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/categories:
    get:
      summary: 모든 카테고리 조회
      description: 모든 카테고리를 조회합니다.
      responses:
        '200':
          description: 카테고리 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array

components:
  schemas:
    Memo:
      type: object
      properties:
        id:
          type: string
          description: 메모의 고유 ID
        title:
          type: string
          description: 메모의 제목
        content:
          type: string
          description: 메모의 내용
        category:
          type: string
          description: 메모의 카테고리
      required:
        - id
        - title
        - content
    MemoInput:
      type: object
      properties:
        title:
          type: string
          description: 메모의 제목
        content:
          type: string
          description: 메모의 내용
        category:
          type: string
          description: 메모의 카테고리
      required:
        - title
        - content
        - category
    CategoryInput:
      type: object
      properties:
        category:
          type: string
          description: 카테고리 이름
      required:
        - category
    Error:
      type: object
      properties:
        error:
          type: string
          description: 에러 메시지
      required:
        - error